requirements.txt
streamlit
groq
PyPDF2
pdfplumber
python-dotenv 
